Change count to uint8_t
Change count for Sensirion I2C HAL from uint16_t to unit8_t.
We never have a count that requires 16-bit integers, therefore we
can safely change it to 8-bit.
This avoids compiler warnings on some platforms.

+++ i2c/sample-implementations/zephyr_user_space/sensirion_i2c_hal.c
@@ -91,7 +91,7 @@ void sensirion_i2c_hal_free(void) {
  * @param count   number of bytes to read from I2C and store in the buffer
  * @returns 0 on success, error code otherwise
  */
-int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint16_t count) {
+int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint8_t count) {
     return i2c_read(i2c_dev, data, count, address);
 }
 
@@ -107,7 +107,7 @@ int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint16_t count) {
  * @returns 0 on success, error code otherwise
  */
 int8_t sensirion_i2c_hal_write(uint8_t address, const uint8_t* data,
-                               uint16_t count) {
+                               uint8_t count) {
     return i2c_write(i2c_dev, data, count, address);
 }
